requirements.txt
python-telegram-bot
pyvista
numpy
matplotlib
pillow
